+++ askbot/doc/source/changelog.rst
@@ -3,6 +3,7 @@ Changes in Askbot
 
 Development master branch (only on github)
 ------------------------------------------
+* Allow custom destination url under the logo
 * Option to allow asking without registration (Egil Moeller)
 * Implemented Mozilla Persona authentication
 * Allowed custom providers of gravatar service (michas2)